Iterators over parallel random-number seeds.

The iRNGStream creates a sequence of random number seeds that are very "far apart" (2^127 steps) in the overall random number sequence, so that each can be used to make a parallel, psudo-independent random iterator. This uses nextRNGStream and the "L'Ecuyer-CMRG" generator (for more details on this mechanism, see vignette("parallel", package="parallel").)

Usage

iRNGStream(seed)

Arguments

seed

Either a single number to be passed to set.seed or a vector to be passed to nextRNGStream or nextRNGSubStream.

Details

iRNGSubStream creates seeds that are somewhat less far apart (2^76 steps), which can be used as "substream" seeds

Originally from the itertools package.

See also

Examples

global.seed <- .Random.seed
#> Error in eval(expr, envir, enclos): object '.Random.seed' not found

rng.seeds <- iRNGStream(313)
print(nextOr(rng.seeds))
#> [1]       10407  -148706266   816896415  -423200657    64647451 -1536586263
#> [7]    64930425
print(nextOr(rng.seeds))
#> [1]       10407   147611069   397195254   280483519 -1571776966 -1788649533
#> [7]  -335749277

# create three pseudo-independent and
# reproducible random number generators
it1 <- isample(c(0, 1), 1, seed=nextOr(rng.seeds))
it2 <- isample(c(0, 1), 1, seed=nextOr(rng.seeds))
it3 <- isample(c(0, 1), 1, seed=nextOr(rng.seeds))

.Random.seed == global.seed
#> Error in eval(expr, envir, enclos): object '.Random.seed' not found
take(it1, 5, "numeric") # 0 0 0 1 1
#> [1] 0 1 1 0 0
take(it2, 5, "numeric") # 0 1 1 1 1
#> [1] 1 0 1 0 1
take(it3, 5, "numeric") # 1 1 1 0 0
#> [1] 0 0 1 1 0

# none of this affects the global seed
global.seed == .Random.seed
#> Error in eval(expr, envir, enclos): object 'global.seed' not found

if (FALSE) {
library(foreach)
foreach(1:3, rseed=iRNGSubStream(1970), .combine='c') %dopar% {
  RNGkind("L'Ecuyer-CMRG") # would be better to initialize workers only once
  assign('.Random.seed', rseed, pos=.GlobalEnv)
  runif(1)
}
}