i_apply {iterors}	R Documentation

Apply a function to each element of an iterator.

Description

i_apply(obj, f) returns the iteror that applies f to each element of the given iterable obj. It is an iterator equivalent of lapply.

Usage

i_apply(obj, f, ...)

Arguments

obj
an iterable.

f
a function

...
Additional arguments will be passed along to f

Value

An iteror.

See Also

To apply a function of multiple arguments to multiple iterators, see i_map. To split an array over margins (like iterators::i_apply use iteror(obj, by=MARGIN

[Package iterors version 1.0.1 Index]