i_readdf {iterors}	R Documentation

Create an iterator to read data frames from files

Description

Create an iterator to read data frames from files.

Usage

i_readdf(filenames, n, start = 1, col.names, chunkSize = 1000)

Arguments

filenames
Names of files contains column data.

n
Maximum number of elements to read from each column file.

start
Element to start reading from.

col.names
Names of the columns.

chunkSize
Number of rows to read at a time.

Details

Originally from the itertools package.

Value

An iteror yielding data.frame objects with up to n rows.

[Package iterors version 1.0.1 Index]