nextOr {iterors}	R Documentation

Retreive the next element from an iteror.

Description

Retreive the next element from an iteror.

Usage

nextOr(obj, or, ...)

Arguments

obj
An iteror.

or
If the iterator has reached its end, this argument will be forced and returned.

...
Other arguments may be used by specific iterors.

Value

Either the next value of iteror, or the value of or.

[Package iterors version 1.0.1 Index]