quantify {iterors}	R Documentation

Count the number of times an iterable object is TRUE

Description

Returns the number of elements from an iterable object that evaluate to TRUE.

Usage

quantify(obj, ...)

Arguments

obj
an iterable object

...
further arguments passed to iteror.

Value

the number of TRUE elements

See Also

reduce

Examples

it <- iteror(c(TRUE, FALSE, TRUE))
quantify(it) # 2

set.seed(42)
x <- sample(c(TRUE, FALSE), size=10, replace=TRUE)
quantify(x) # Equivalent to sum(x)

[Package iterors version 1.0.1 Index]