Efficient, compact iteration.

it <- iteror(obj, ...) is a generic constructor that creates objects of class "iteror" from its input. An iteror outputs a single element of a sequence each time you call nextOr(it). Different iteror methods exist for different data types and may take different optional arguments as listed in this page.

Usage

iteror(obj, ...)

# S3 method for default
iteror(obj, ..., recycle = FALSE, chunkSize, chunks)

# S3 method for connection
iteror(obj, ...)

# S3 method for array
iteror(
  obj,
  ...,
  by = c("cell", "row", "column"),
  chunkSize,
  chunks,
  recycle = FALSE,
  drop = FALSE,
  rowMajor = TRUE
)

# S3 method for matrix
iteror(
  obj,
  ...,
  by = c("cell", "row", "column"),
  chunkSize,
  chunks,
  recycle = FALSE,
  drop = FALSE,
  rowMajor = TRUE
)

# S3 method for data.frame
iteror(obj, ..., recycle = FALSE, chunkSize, chunks, by = c("column", "row"))

Arguments

obj

An object to iterate with.

...

Different iteror methods may take differetn options depending on the class of obj.#' @return an object of classes 'iteror' and 'iter'.

recycle

a boolean describing whether the iterator should reset after running through all its values.

chunkSize

the number of elements of by to return with each call to nextOr.

chunks

Split the input into this many chunks.

by

how to split up an array or data frame. Can be "cell", "row", "column", or numeric dimensions.

drop

Whether to drop the array dimensions enumerated over.

rowMajor

If TRUE, will return slices in order with the first indices varying fastest (same as in ienumerate).

Details

When called, an iterator may either return a new value or stop. The way an iteror signals a stop is that it does whatever you write in the argument or. For instance you can write or=break to exit a loop.Summing over an iteror this way looks like:

sum <- 0
it <- iteror(iseq(0, 100, 7))
repeat {
  sum <- sum + nextOr(it, break)
}

Another way to use the "or" argument is to give it a sigil value; that is, a special value that will be interpreted as end of iteration. If the result of calling nextOr is identical() to the sigil value you provided, then you know the iterator has ended. This pattern looks like:

sum <- 0
stopped <- new.env()
it <- iteror(iseq(0, 100, 7))
repeat {
  val <- nextOr(it, stopped)
  if (identical(val, stopped)) break
  sum <- sum + val
}

(Note the use of new.env(). In R it is commonplace to use NULL or NA in the role of a sigil, but that only works until you have an iterator that needs to yield NULL itself. A safer alternative is to use a one-shot sigil value; new.env() works for this, as it constructs an object that is not identical to any other object in the R session.)

Note that iteror objects are simply functions with a class attribute attached, and all nextOr.iteror does is call the function. So if you were in the mood, you could skip calling nextOr through S3 dispatch and call the function directly. If you take this approach, make sure you have called iteror() to ensure that you have a true iteror object.

sum <- 0
it <- iteror(iseq(0, 100, 7))
repeat sum <- sum + it(or=break)
sum
#> [1] 735

To create iterors with custom-defined behavior, see iteror.function.

Examples

l <- iteror(letters, chunkSize=7)
as.list(l)
#> [[1]]
#> [1] "a" "b" "c" "d" "e" "f" "g"
#> 
#> [[2]]
#> [1] "h" "i" "j" "k" "l" "m" "n"
#> 
#> [[3]]
#> [1] "o" "p" "q" "r" "s" "t" "u"
#> 
#> [[4]]
#> [1] "v" "w" "x" "y" "z"
#> 

a <- array(1:8, c(2, 2, 2))

# iterate over all the slices
it <- iteror(a, by=3)
as.list(it)
#> [[1]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    2    4
#> 
#> 
#> [[2]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    5    7
#> [2,]    6    8
#> 
#> 

# iterate over all the columns of each slice
it <- iteror(a, by=c(2, 3))
as.list(it)
#> [[1]]
#> , , 1
#> 
#>      [,1]
#> [1,]    1
#> [2,]    2
#> 
#> 
#> [[2]]
#> , , 1
#> 
#>      [,1]
#> [1,]    3
#> [2,]    4
#> 
#> 
#> [[3]]
#> , , 1
#> 
#>      [,1]
#> [1,]    5
#> [2,]    6
#> 
#> 
#> [[4]]
#> , , 1
#> 
#>      [,1]
#> [1,]    7
#> [2,]    8
#> 
#> 

# iterate over all the rows of each slice
it <- iteror(a, by=c(1, 3))
as.list(it)
#> [[1]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    1    3
#> 
#> 
#> [[2]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    2    4
#> 
#> 
#> [[3]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    5    7
#> 
#> 
#> [[4]]
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]    6    8
#> 
#>